Fix non lowercase distinfo_name on Windows
Fixes #11

setuptools.package_index.PackageIndex used by zc.buildout will have
called, pkg_resources.normalize_path on the location, which in turn
calls os.path.normcase on it, thus lowercasing it on Windows. This means
that when we try to create a wheel.install.WheelFile using the location,
the WheelFile.distinfo_name is calculated from this altered path, so the
WheelFile.record_name uses an incorrectly lowercased version of the
dist-info directory to look for RECORD and as zipfile uses
case-sensitive paths this will fail. Fix by using the exact same case we
have on the filesystem.